requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-whisper==20250625
packaging==25.0
pandas==2.3.0
pillow==11.3.0
proglog==0.1.12
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
speechbrain==1.0.3
sympy==1.14.0
tabulate==0.9.0
thefuzz==0.22.1
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
triton==3.3.1
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
